Add scenario count to check for accidental disabling of tests
