+++ app_spec/test/index.js
@@ -4,6 +4,11 @@ const tape = require('tape')
 const { Orchestrator, tapeExecutor, backwardCompatibilityMiddleware } = require('@holochain/try-o-rama')
 const spawnConductor = require('./spawn_conductors')
 
+// This constant serves as a check that we haven't accidentally disabled scenario tests.
+// Try to keep this number as close as possible to the actual number of scenario tests.
+// (But never over)
+const MIN_EXPECTED_SCENARIOS = 50
+
 process.on('unhandledRejection', error => {
   // Will print "unhandledRejection err is not defined"
   console.error('got unhandledRejection:', error);
@@ -48,12 +53,34 @@ const orchestratorMultiDna = new Orchestrator({
   callbacksPort: 8888,
 })
 
-require('./regressions')(orchestratorSimple.registerScenario)
-require('./test')(orchestratorSimple.registerScenario)
-require('./multi-dna')(orchestratorMultiDna.registerScenario)
-require('./validate-agent-test')(dnaPath)
+const orchestratorValidateAgent = new Orchestrator({
+  conductors: {
+    valid_agent: { instances: { app: dna } },
+    reject_agent: { instances: { app: dna } },
+  },
+  debugLog: false,
+  executor: tapeExecutor(require('tape')),
+  middleware: backwardCompatibilityMiddleware,
+})
 
-const run = async () => {
+const registerAllScenarios = () => {
+  let numRegistered = 0
+
+  const registerer = orchestrator => (...info) => {
+    numRegistered += 1
+    return orchestrator.registerScenario(...info)
+  }
+
+  require('./regressions')(registerer(orchestratorSimple))
+  require('./test')(registerer(orchestratorSimple))
+  require('./multi-dna')(registerer(orchestratorMultiDna))
+  require('./validate-agent-test')(registerer(orchestratorValidateAgent))
+
+  return numRegistered
+}
+
+
+const runTests1 = async () => {
   const alice = await spawnConductor('alice', 3000)
   await orchestratorSimple.registerConductor({name: 'alice', url: 'http://0.0.0.0:3000'})
   const bob = await spawnConductor('bob', 4000)
@@ -79,7 +106,37 @@ const run = async () => {
   //await orchestratorMultiDna.registerConductor({name: 'conductor', url: 'http://0.0.0.0:6000'})
   //await orchestratorMultiDna.run()
   //conductor.kill()
+}
 
+const runValidationTests = async () => {
+  const valid_agent = await spawnConductor('valid_agent', 3000)
+  await orchestrator.registerConductor({name: 'valid_agent', url: 'http://0.0.0.0:3000'})
+  const reject_agent = await spawnConductor('reject_agent', 4000)
+  await orchestrator.registerConductor({name: 'reject_agent', url: 'http://0.0.0.0:4000'})
+
+  const delay = ms => new Promise(resolve => setTimeout(resolve, ms))
+  console.log("Waiting for conductors to settle...")
+  await delay(5000)
+  console.log("Ok, starting tests!")
+
+  await orchestrator.run()
+  valid_agent.kill()
+  reject_agent.kill()
+}
+
+const run = async () => {
+  const num = registerAllScenarios()
+
+  // Check to see that we haven't accidentally disabled a bunch of scenarios
+  if (num < MIN_EXPECTED_SCENARIOS) {
+    console.error(`Expected at least ${MIN_EXPECTED_SCENARIOS}, but only ${num} were registered!`)
+    process.exit(1)
+  } else {
+    console.log(`Registered ${num} scenarios (at least ${MIN_EXPECTED_SCENARIOS} were expected)`)
+  }
+
+  await runTests1()
+  await runValidationTests()
   process.exit()
 }
